+++ NEWS.md
@@ -1,3 +1,5 @@
+# jfa 0.5.3
+
 # jfa 0.5.2
 
 - Update the poisson evaluation calculation so that it allows for fractional errors.